Add files via upload
This is a web page to introduce UFL, Gainesville and Florida. You can get access to different information by tab the related tabs.
The click event is processed by Javascript to realize the display and disappearance of the text part. Each box has a small arrow pointing to the text object.
The inner shadow of the box is processed through CSS, and the overflow part is hidden to make the page more beautiful.
This project includes skills of CSS, HTML and Javascript.